Separated action buttons from image carousel.
