+++ .env.oss
@@ -11,9 +11,3 @@ USER_ID=
 USER_ACCESS_TOKEN=
 
 APP_URL=
-FORCE_CLOUDFRONT_URL=
-GEMINI_CLOUDFRONT_URL=
-METAPHYSICS_ENDPOINT=
-NODE_ENV=
-STRIPE_PUBLISHABLE_KEY=
-XAPP_TOKEN=
